+++ Venmo/src/main/java/com/braintreepayments/api/VenmoLauncherCallback.java
@@ -3,7 +3,7 @@ package com.braintreepayments.api;
 /**
  * Used to receive notification that the Venmo payment authorization flow completed
  * Once this is invoked, continue the flow by calling
- * {@link VenmoClient#tokenize(VenmoPaymentAuthResult, VenmoResultCallback)}
+ * {@link VenmoClient#tokenize(VenmoPaymentAuthResult, VenmoTokenizeCallback)}
  */
 public interface VenmoLauncherCallback {