+++ Venmo/src/main/java/com/braintreepayments/api/VenmoPaymentAuthRequestCallback.java
@@ -4,7 +4,7 @@ import androidx.annotation.Nullable;
 
 /**
  * Callback to handle result from
- * {@link VenmoClient#tokenize(VenmoPaymentAuthResult, VenmoResultCallback)}
+ * {@link VenmoClient#tokenize(VenmoPaymentAuthResult, VenmoTokenizeCallback)}
  */
 public interface VenmoPaymentAuthRequestCallback {
     void onPaymentAuthRequest(@Nullable VenmoPaymentAuthRequest venmoPaymentAuthRequest,